+++ src/Serilog.Seq.Test/SerilogSeqSinkHelperTest.cs
@@ -38,7 +38,7 @@ namespace Serilog.Seq.Test
 			var seqHost = "http://nevermind";
 			_fixture.Inject(seqHost);
 			var seqServerMock = _fixture.Create<Mock<SeqServer>>();
-			seqServerMock.Setup(server => server.RegisterApplicationAsync(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<CancellationToken>())).Returns(Task.CompletedTask);
+			seqServerMock.Setup(server => server.RegisterApplicationAsync(It.IsAny<SeqServerApplicationInformation>(), It.IsAny<string>(), It.IsAny<CancellationToken>())).Returns(Task.CompletedTask);
 			var seqServer = seqServerMock.Object;
 
 			// Act
@@ -55,7 +55,7 @@ namespace Serilog.Seq.Test
 			var seqHost = "http://nevermind";
 			_fixture.Inject(seqHost);
 			var seqServerMock = _fixture.Create<Mock<SeqServer>>();
-			seqServerMock.Setup(server => server.RegisterApplicationAsync(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<CancellationToken>())).Throws(_fixture.Create<SeqServerApplicationRegisterException>());
+			seqServerMock.Setup(server => server.RegisterApplicationAsync(It.IsAny<SeqServerApplicationInformation>(), It.IsAny<string>(), It.IsAny<CancellationToken>())).Throws(_fixture.Create<SeqServerApplicationRegisterException>());
 			var seqServer = seqServerMock.Object;
 
 			// Act
@@ -72,7 +72,7 @@ namespace Serilog.Seq.Test
 			var seqHost = "http://nevermind";
 			_fixture.Inject(seqHost);
 			var seqServerMock = _fixture.Create<Mock<SeqServer>>();
-			seqServerMock.Setup(server => server.RegisterApplicationAsync(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<CancellationToken>())).Throws(_fixture.Create<SeqServerApplicationRegisterException>());
+			seqServerMock.Setup(server => server.RegisterApplicationAsync(It.IsAny<SeqServerApplicationInformation>(), It.IsAny<string>(), It.IsAny<CancellationToken>())).Throws(_fixture.Create<SeqServerApplicationRegisterException>());
 			var seqServer = seqServerMock.Object;
 
 			// Act