+++ src/Serilog.Test/LoggerSettingsConfigurationExtensionsTest.cs
@@ -1,5 +1,6 @@
 using System;
 using System.IO;
+using System.Linq;
 using System.Reflection;
 using System.Runtime.InteropServices;
 using System.Text;
@@ -13,6 +14,23 @@ namespace Serilog.Test
 {
 	public class LoggerSettingsConfigurationExtensionsTest
 	{
+		public string ValidConfigurationContent => @"
+		{
+		  ""Serilog"": {
+		    ""Using"":  [ ""Serilog.Sinks.Debug"", ""Serilog.Sinks.Console"" ],
+		    ""MinimumLevel"": ""Fatal"",
+		    ""WriteTo"": [
+		      { ""Name"": ""Debug"" },
+		      { ""Name"": ""Console"" },
+		    ],
+		    ""Enrich"": [ ""FromLogContext"", ""WithMachineName"", ""WithThreadId"" ],
+		    ""Properties"": {
+		        ""Application"": ""Sample""
+		    }
+		  }
+		}
+		";
+
 		[SetUp]
 		public void Setup() { }
 
@@ -43,22 +61,7 @@ namespace Serilog.Test
 		{
 			// Arrange
 			using var testDirectory = new TestDirectory();
-			var contentBuilder = new StringBuilder();
-			contentBuilder.AppendLine(@"{");
-			contentBuilder.AppendLine(@"  ""Serilog"": {");
-			contentBuilder.AppendLine(@"    ""Using"":  [ ""Serilog.Sinks.Debug"", ""Serilog.Sinks.Console"" ],");
-			contentBuilder.AppendLine(@"    ""MinimumLevel"": ""Fatal"",");
-			contentBuilder.AppendLine(@"    ""WriteTo"": [");
-			contentBuilder.AppendLine(@"      { ""Name"": ""Debug"" },");
-			contentBuilder.AppendLine(@"      { ""Name"": ""Console"" },");
-			contentBuilder.AppendLine(@"    ],");
-			contentBuilder.AppendLine(@"    ""Enrich"": [ ""FromLogContext"", ""WithMachineName"", ""WithThreadId"" ],");
-			contentBuilder.AppendLine(@"    ""Properties"": {");
-			contentBuilder.AppendLine(@"        ""Application"": ""Sample""");
-			contentBuilder.AppendLine(@"    }");
-			contentBuilder.AppendLine(@"  }");
-			contentBuilder.AppendLine(@"}");
-			var configurationFile = testDirectory.CreateFile("serilog.config", contentBuilder.ToString());
+			var configurationFile = testDirectory.CreateFile("serilog.config", this.ValidConfigurationContent);
 
 			// Act
 			var configuration = new LoggerConfiguration().ReadFrom.JsonFile(configurationFile);
@@ -77,6 +80,34 @@ namespace Serilog.Test
 			var sinksField = type?.GetField("_sinks", BindingFlags.Instance | BindingFlags.NonPublic);
 			var sinks = (Array) sinksField?.GetValue(aggregateSink);
 			Assert.That(sinks, Has.Length.EqualTo(2));
+		}
+
+		[Test]
+		public void Check_Configuration_File_Is_Copied_To_Working_Directory()
+		{
+			// Arrange
+			var configFileName = "serilog.config";
+			var applicationDirectory = new DirectoryInfo(AppDomain.CurrentDomain.BaseDirectory);
+			var defaultConfigFile = new FileInfo(Path.Combine(applicationDirectory.FullName, configFileName));
+			if (defaultConfigFile.Exists) defaultConfigFile.Delete();
+			using var testDirectory = new TestDirectory();
+
+			try
+			{
+				// Arrange
+				File.WriteAllText(defaultConfigFile.FullName, this.ValidConfigurationContent);
+				Directory.SetCurrentDirectory(testDirectory.Directory.FullName);
+				
+				// Act + Assert
+				Assert.DoesNotThrow(() => new LoggerConfiguration().ReadFrom.JsonFile(configFileName));
+				Assert.DoesNotThrow(() => testDirectory.Directory.EnumerateFiles().Single(file => file.Name == configFileName));
+			}
+			finally
+			{
+				Directory.SetCurrentDirectory(applicationDirectory.FullName);
+				defaultConfigFile.Refresh();
+				if (defaultConfigFile.Exists) defaultConfigFile.Delete();
+			}
 		}
 	}
 }